refactor(macos): split SessionManager god-object into cohesive extensions
CodeScene flagged SessionManager.swift as a hotspot (Code Health 5.86/10,
1456+ LOC, 68 revisions, steady friction) — issue #807. Root cause: it was
a god object bundling ~10 unrelated concerns (local WebSocket, relay
WebSocket + source reconciliation, REST hydration/consent, history-bar
wire decoding, apiGroups tree patching, notifications, persisted session/
group ordering, session actions, debug dump), so any lifecycle/state-
machine change touched the same 2000+ line file as everything else.

The three-state (working/waiting/ready) invariant itself is already
centralized in SessionState.State — no duplication found there to fix.

Split the class's behavior into 9 `SessionManager+*.swift` extension
files by concern, each with its own MARK header explaining its slice.
All stored state, init, and deinit stay in the trimmed core file (305
lines, down from 2053) since Swift requires those in the primary
declaration. No behavior change — verified by diffing the full symbol
set (func/struct/class/enum and @Published counts) between the original
file and the sum of the new files, plus a green build + test run.

Mechanical consequence: members previously `private`/`private(set)` that
are now called across file boundaries had to become internal (module-
scoped) since Swift's `private` is file-lexical. This doesn't change any
externally observable behavior — the app is a single compiled target.
